Quality of life improvements to `konfigure kustomizepatch` (#88)
* increase timeout when calling source-controller

15s changed to 60s.
Helps on clusters with slower network. Sometimes 15s was not enough to
download the packed repository. We do not care about longer waits, as
long as the download is successful. Since repositories are cached, we
do not expect unnecessary calls to be made.

* add logging to kustomizepatch

Increased verbosity will help us debug issues when konfigure is running
as a kustomize/kustomize-controller plugin. The log messages will blend
with kustomize-controller's output.

* update CHANGELOG