Define includeSelf template rendering function + consider missing default templates empty (#218)
* Define includeSelf template rendering function

With config split we move the original include folder of the giantswarm/config
repository to be a submodule in customer specific config repositories.

In order to support customer specific include files we add support for the
include-self directory. Includes from this folder can be imported by the new
includeSelf function.

Keeping include and adding includeSelf is better for error reporting and
making it clear where the include is coming from instead of supporting fallbacks
to find the file and masking error messages on problems.